+++ apps/graduate/src/shared/components/StudentAddModal/StudentAddMultipleModel/StudentAddMultiple.tsx
@@ -1,79 +1,42 @@
-import { Button, Typography, Upload, message } from 'antd';
-import { useEffect, useState } from 'react';
+import { Button, Typography, Upload } from 'antd';
 
-import { PROFESSORS } from '~/shared/constants/professors';
-
+import type { BulkUploadRow } from '../types';
 import BulkPreviewTable from './BulkPreviewTable';
-import { createProfessorMap, parseCsv, parseXlsx } from './parsers';
-import type { BulkUploadRow, InvalidRow } from '../types';
+import { useBulkUpload } from './useBulkUpload';
 
 type Props = {
   open?: boolean;
   onBulkSubmit?: (rows: BulkUploadRow[]) => void | Promise<void>;
 };
 
-const PROFESSOR_NAME_TO_ID = createProfessorMap(PROFESSORS);
+export default function StudentAddMultiple({
+  open = false,
+  onBulkSubmit,
+}: Props) {
+  const {
+    fileName,
+    rows,
+    invalidRows,
+    selectedRowKeys,
+    current,
+    pageSize,
+    setSelectedRowKeys,
+    setCurrent,
+    handleFileChange,
+    handleResultAlert,
+  } = useBulkUpload(open);
 
-export default function StudentAddMultiple({ open, onBulkSubmit }: Props) {
-  const [fileName, setFileName] = useState('');
-  const [rows, setRows] = useState<BulkUploadRow[]>([]);
-  const [invalidRows, setInvalidRows] = useState<InvalidRow[]>([]);
-  const [selectedRowKeys, setSelectedRowKeys] = useState<React.Key[]>([]);
-  const pageSize = 7;
-  const [current, setCurrent] = useState(1);
-
-  const onSelectionChange = (keys: React.Key[]) => setSelectedRowKeys(keys);
-
-  useEffect(() => {
-    if (!open) return;
-    setFileName('');
-    setRows([]);
-    setInvalidRows([]);
-    setSelectedRowKeys([]);
-    setCurrent(1);
-  }, [open]);
-
-  const onFileChange = async (file: File) => {
-    setFileName(file.name);
-    const name = file.name.toLowerCase();
-
-    let result: { valid: BulkUploadRow[]; invalid: InvalidRow[] } = {
-      valid: [],
-      invalid: [],
-    };
-
-    if (name.endsWith('.csv')) {
-      const text = await file.text();
-      result = parseCsv(text, PROFESSOR_NAME_TO_ID);
-    } else if (name.endsWith('.xlsx') || name.endsWith('.xls')) {
-      result = await parseXlsx(file, PROFESSOR_NAME_TO_ID);
-    } else {
-      message.warning('CSV 또는 XLSX 파일만 업로드 가능합니다.');
-      setRows([]);
-      setInvalidRows([]);
-      return false;
-    }
-
-    setRows(result.valid);
-    setInvalidRows(result.invalid);
-    setCurrent(1);
-    setSelectedRowKeys([]);
-    return false;
+  const handleSelectSubmit = async () => {
+    const selected = rows.filter(r => selectedRowKeys.includes(r.key));
+    await onBulkSubmit?.(selected);
+    handleResultAlert(selected.length);
   };
 
-  const showResultModal = (addedCount: number) => {
-    const invalidMsg = invalidRows
-      .map(r => `${r.name}[${r.studentNo}] : ${r.reason}`)
-      .join('\n');
-
-    alert(
-      `${addedCount}명의 인원이 추가되었습니다.\n` +
-        `${
-          invalidRows.length > 0
-            ? `제외된 인원들 ${invalidRows.length}명 :\n` + invalidMsg
-            : ''
-        }`,
-    );
+  const handleAllSubmit = async () => {
+    if (rows.length > 0) {
+      await onBulkSubmit?.(rows);
+    }
+    handleResultAlert(rows.length);
   };
 
   return (
@@ -92,9 +55,9 @@ export default function StudentAddMultiple({ open, onBulkSubmit }: Props) {
           rows={rows}
           pageSize={pageSize}
           current={current}
-          onPageChange={p => setCurrent(p)}
+          onPageChange={setCurrent}
           selectedRowKeys={selectedRowKeys}
-          onSelectionChange={onSelectionChange}
+          onSelectionChange={setSelectedRowKeys}
         />
       </div>
 
@@ -102,7 +65,7 @@ export default function StudentAddMultiple({ open, onBulkSubmit }: Props) {
         style={{ display: 'flex', alignItems: 'center', gap: 8, marginTop: 12 }}
       >
         <Upload
-          beforeUpload={onFileChange}
+          beforeUpload={handleFileChange}
           showUploadList={false}
           accept='.csv,.xlsx,.xls'
         >
@@ -114,23 +77,14 @@ export default function StudentAddMultiple({ open, onBulkSubmit }: Props) {
         <div style={{ flex: 1 }} />
         <Button
           disabled={!rows.length || !selectedRowKeys.length}
-          onClick={async () => {
-            const selected = rows.filter(r => selectedRowKeys.includes(r.key));
-            await onBulkSubmit?.(selected);
-            showResultModal(selected.length);
-          }}
+          onClick={handleSelectSubmit}
         >
           선택 등록
         </Button>
         <Button
           type='primary'
           disabled={!rows.length && !invalidRows.length}
-          onClick={async () => {
-            if (rows.length > 0) {
-              await onBulkSubmit?.(rows);
-            }
-            showResultModal(rows.length);
-          }}
+          onClick={handleAllSubmit}
         >
           일괄 등록
         </Button>
